fix: 重複する内容をファイル作成で統一
