+++ frontend/src/shared/components/register-form.ts
@@ -1,5 +1,6 @@
 import { AuthService } from "../services/auth-service";
 import type { CreateUserRequest, PublicUser } from "../types/user";
+import { setupPasswordToggles } from "../utils/password-toggle-utils";
 
 const eyeIconUrl = new URL("../../../images/icon/eye_icon.png", import.meta.url)
   .href;
@@ -117,7 +118,7 @@ export class RegisterForm {
     `;
 
     this.attachEventListeners();
-    this.setupPasswordToggles();
+    setupPasswordToggles(this.container, eyeIconUrl);
   }
 
   private attachEventListeners(): void {
@@ -145,54 +146,6 @@ export class RegisterForm {
     showHomeBtn.addEventListener("click", () => this.onShowHome(), {
       signal: this.abortController.signal,
     });
-  }
-
-  private setupPasswordToggles(): void {
-    const toggles =
-      this.container.querySelectorAll<HTMLButtonElement>(".password-toggle");
-
-    toggles.forEach((toggle) => {
-      const targetId = toggle.dataset.target;
-      const input = targetId
-        ? this.container.querySelector<HTMLInputElement>(`#${targetId}`)
-        : null;
-
-      if (!input) {
-        console.warn("Password toggle target not found for register form.");
-        return;
-      }
-
-      const applyState = (visible: boolean) => {
-        input.type = visible ? "text" : "password";
-        toggle.dataset.visible = String(visible);
-        toggle.setAttribute(
-          "aria-label",
-          visible ? "Hide password" : "Show password",
-        );
-        toggle.innerHTML = this.getPasswordToggleIcon(visible);
-      };
-
-      const initialVisible = toggle.dataset.visible === "true";
-      applyState(initialVisible);
-
-      toggle.addEventListener("click", () => {
-        const isVisible = toggle.dataset.visible === "true";
-        applyState(!isVisible);
-      });
-    });
-  }
-
-  private getPasswordToggleIcon(isVisible: boolean): string {
-    const slashMarkup = isVisible
-      ? ""
-      : `<span class="absolute block" style="width: 1.35rem; height: 2px; background-color: currentColor; transform: rotate(45deg); border-radius: 9999px;"></span>`;
-
-    return `
-      <span class="relative inline-flex h-5 w-5 items-center justify-center">
-        <img src="${eyeIconUrl}" alt="" class="pointer-events-none h-5 w-5 object-contain" />
-        ${slashMarkup}
-      </span>
-    `;
   }
 
   private async handleSubmit(event: Event): Promise<void> {
